+++ tools/ndctl/cxl/volatile/vmem.sh
@@ -2,13 +2,17 @@
 set -e
 # commit d5a21a914482 ("cxl.sh: multi memdevs to on region")
 
-# create region0, commit b3c049d89aa9 ("cxl: cxl.sh: create-region")
-# will create
+# Create region0, commit b3c049d89aa9 ("cxl: cxl.sh: create-region")
+# create:
 # - /sys/bus/cxl/devices/region0/dax_region
 # - /sys/devices/platform/ACPI0017:00/root0/decoder0.0/region0/dax_region0/dax0.0
 # - /sys/bus/dax/devices/dax0.0
 # - /sys/bus/dax/drivers/kmem/dax0.0
 # - /dev/dax0.0 [Character Device], commit 1227030c1d3f ("cxl.sh: region: character device /dev/dax0.0")
+#
+# Write:
+# - "region0" to /sys/bus/cxl/devices/root0/decoder0.0/create_ram_region
+# - 1024M     to /sys/bus/cxl/devices/root0/decoder0.0/region0/size
 sudo cxl create-region --decoder decoder0.0 --size 4096M --type ram --memdevs mem0 mem1 mem2 mem3
 # or only use one way
 sudo cxl create-region --decoder decoder0.0 --size 1024M --type ram --memdevs mem0