+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- cxl: create region with `/sys` fs;
+- sysfs: add `/sys/bus/cxl` and `/sys/devices/platform/ACPI0017:00/` subdir;
 - python: test locale/getlocale.py;
 - scripts/rootfs: add debian.sh;
 - version: config.json: add OpenIPMI;